+++ lib/active_record_query_counter.rb
@@ -233,12 +233,8 @@ module ActiveRecordQueryCounter
     # @param connection_class [Class] the connection adapter class to extend
     # @return [void]
     def enable!(connection_class)
-      unless connection_class.include?(ConnectionAdapterExtension)
-        connection_class.prepend(ConnectionAdapterExtension)
-      end
-      unless ActiveRecord::ConnectionAdapters::TransactionManager.include?(TransactionManagerExtension)
-        ActiveRecord::ConnectionAdapters::TransactionManager.prepend(TransactionManagerExtension)
-      end
+      ConnectionAdapterExtension.inject(connection_class)
+      TransactionManagerExtension.inject(ActiveRecord::ConnectionAdapters::TransactionManager)
 
       @cache_subscription ||= ActiveSupport::Notifications.subscribe("sql.active_record") do |_name, _start_time, _end_time, _id, payload|
         if payload[:cached]