+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.0
+
+### Added
+
+- Added `ActiveRecordQueryCounter.disable` method to allow disabling query counting behavior within a block.
+- Rails 7.1 compatibility.
+
 ## 2.1.0
 
 ### Added